refactor: extract algolia params builders to new files in a new corresponding folder
