+++ src/libs/algolia/fetchAlgolia/buildAlgoliaParameters/buildGeolocationParameter.ts
@@ -2,71 +2,7 @@ import { LocationType } from 'features/search/enums'
 import { PartialSearchState } from 'features/search/types'
 import { GeoCoordinates } from 'libs/geolocation'
 
-import { RADIUS_FILTERS } from '../enums'
-
-import { buildFacetFilters } from './fetchAlgolia.facetFilters'
-import { buildNumericFilters } from './fetchAlgolia.numericFilters'
-
-// We don't use all the fields indexed. Simply retrieve the one we use.
-// see SearchHit
-export const offerAttributesToRetrieve = [
-  'offer.dates',
-  'offer.isDigital',
-  'offer.isDuo',
-  'offer.isEducational',
-  'offer.name',
-  'offer.prices',
-  'offer.subcategoryId',
-  'offer.thumbUrl',
-  'objectID',
-  '_geoloc',
-]
-
-export const buildOfferSearchParameters = (
-  {
-    beginningDatetime = null,
-    date = null,
-    endingDatetime = null,
-    locationFilter,
-    offerCategories = [],
-    offerSubcategories = [],
-    objectIds = [],
-    offerIsDuo = false,
-    offerIsFree = false,
-    offerIsNew = false,
-    offerTypes = {
-      isDigital: false,
-      isEvent: false,
-      isThing: false,
-    },
-    priceRange = null,
-    timeRange = null,
-    tags = [],
-  }: PartialSearchState & { objectIds?: string[] },
-  userLocation: GeoCoordinates | null,
-  isUserUnderage: boolean
-) => ({
-  ...buildFacetFilters({
-    locationFilter,
-    offerCategories,
-    offerSubcategories,
-    objectIds,
-    offerTypes,
-    offerIsDuo,
-    tags,
-    isUserUnderage,
-  }),
-  ...buildNumericFilters({
-    beginningDatetime,
-    date,
-    endingDatetime,
-    offerIsFree,
-    offerIsNew,
-    priceRange,
-    timeRange,
-  }),
-  ...buildGeolocationParameter(locationFilter, userLocation),
-})
+import { RADIUS_FILTERS } from '../../enums'
 
 export const buildGeolocationParameter = (
   locationFilter: PartialSearchState['locationFilter'],
